requirements.txt
imageio==2.8.0
pafy==0.5.5
numpy==1.18.4
opencv-python==4.2.0.34
pytesseract==0.3.4
pillow==6.2.2
nltk==3.5
youtube_dl==2020.5.8    
moviepy==1.0.3
